\name{assortedViz-package}
\alias{assortedViz-package}
\alias{assortedViz}
\docType{package}
\title{
\packageTitle{assortedViz}
}
\description{
\packageDescription{assortedViz}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{assortedViz}
\packageIndices{assortedViz}
}
\author{
\packageAuthor{assortedViz}

Maintainer: \packageMaintainer{assortedViz}
}
